requirements.txt
certifi==2026.2.25
Pillow>=10.0.0
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.13.4
cryptography==46.0.5
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
ply==3.11
pyasn1==0.6.2
pycparser==3.0
pycryptodomex==3.23.0
Pygments==2.19.2
pysmi==1.6.3
pysmi-lextudio==1.4.3
pysnmp==7.1.22
pytest==9.0.2
pytest-cov==4.1.0
requests==2.32.5
urllib3==2.6.3
